fixed NTP protocol monitoring
git-svn-id: svn://svn.zabbix.com/trunk@3043 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ include/sysinfo.h
@@ -54,7 +54,7 @@ void    add_user_parameter(char *key,char *command);
 void	test_parameters(void);
 void	test_parameter(char* key);
 
-int     check_ntp(char *host, int port, int *value_int);
+int     check_ntp(char *host, unsigned short port, int *value_int);
 
 int     get_stat(const char *key, unsigned flags, AGENT_RESULT *result);
